% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_miss}
\alias{check_miss}
\title{Function to count missing data in input dataset}
\usage{
check_miss(data, ...)
}
\arguments{
\item{data}{inputdataset.}

\item{...}{arguments pass to dplyr (in progress).}
}
\value{
data.frame with number of NA (N_miss), number of 0 length characters (N_0_char), number of unique values (N_unique) and number of observations (N).
}
\description{
Function that returns information bout missing values,
characters that have 0 length as well as number of
unique values and numer of observations
}
\examples{
check_miss(mtcars)
}

