% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lb}
\alias{lb}
\alias{lb_d}
\alias{lb_l}
\title{Function for loading packages}
\usage{
lb_l(...)

lb_d(name, ...)
}
\arguments{
\item{...}{arguments passed to library and detach}

\item{name}{name of package to be detached.}
}
\description{
Function for loading packages
}
\examples{
lb_l(dplyr,stringr)
lb_d(dplyr)
}

