% Generated by roxygen2 (4.0.2): do not edit by hand
\name{str_dummy}
\alias{str_dummy}
\title{Function to create dummy variables from strings}
\usage{
str_dummy(var, strs, ...)
}
\arguments{
\item{var}{character vector that will be coverted to dummy variables.}

\item{strs}{charcter vector that contains values that will create columns.}

\item{...}{to be passed to grepl function}
}
\value{
returns a data.frame with 0-1 values.
}
\description{
Function to create dummy variables from strings
}
\examples{
v1 <- c('A,B','A,B,C','C','B,A','C,A')
str_dummy(v1,c('A','B','C'))
}

